+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   - Drawn entities can be rotated.
   - Dimensions of rectangles can now be set using forms when selected.
   - Added copy function for drawn models.
+  - Added options section to draw overview.
 ### Changed
 - Modeler3D:
   - Polygon and line highlighting has been changed